% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BloodPressureCategory.R
\name{BloodPressureCategory}
\alias{BloodPressureCategory}
\title{Blood Pressure Category Function}
\usage{
BloodPressureCategory()
}
\value{
Prints the category of blood pressure based on the systolic and diastolic inputs.
}
\description{
This function prompts the user to input their systolic and diastolic blood pressure readings and then categorizes the blood pressure based on the provided values.
}
\details{
The categories are defined as follows:
\itemize{
\item \strong{Normal:} Systolic < 120 mmHg and Diastolic < 80 mmHg
\item \strong{Elevated:} Systolic 120-129 mmHg and Diastolic < 80 mmHg
\item \strong{High blood pressure (Hypertension) Stage 1:} Systolic 130-139 mmHg or Diastolic 80-89 mmHg
\item \strong{High blood pressure (Hypertension) Stage 2:} Systolic 140-180 mmHg or Diastolic ≥ 90 mmHg
\item \strong{Hypertensive crisis:} Systolic > 180 mmHg or Diastolic > 120 mmHg
}
}
