% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesterol_levels.R
\name{cholesterol_levels}
\alias{cholesterol_levels}
\title{Cholesterol Levels Function}
\usage{
cholesterol_levels()
}
\value{
Prints the category of cholesterol levels based on the inputs.
}
\description{
This function prompts the user to input their age and cholesterol readings (LDL, triglycerides, and HDL) and then categorizes the cholesterol levels based on the provided values.
}
\details{
The cholesterol levels are categorized based on age:
\itemize{
\item \strong{For age <= 19:}
\itemize{
\item Normal cholesterol: Total cholesterol < 170 mg/dL, Non-HDL cholesterol < 120 mg/dL, HDL cholesterol > 45 mg/dL, and LDL cholesterol < 110 mg/dL
\item Borderline high: Total cholesterol 170-199 mg/dL, Non-HDL cholesterol 120-144 mg/dL, HDL cholesterol 50-55 mg/dL, and LDL cholesterol 110-129 mg/dL
\item High: Total cholesterol ≥ 200 mg/dL, Non-HDL cholesterol ≥ 145 mg/dL, HDL cholesterol < 45 mg/dL, and LDL cholesterol ≥ 130 mg/dL
}
\item \strong{For age >= 20:}
\itemize{
\item Normal cholesterol: Total cholesterol < 200 mg/dL, Non-HDL cholesterol < 130 mg/dL, HDL cholesterol ≥ 40 mg/dL, and LDL cholesterol < 100 mg/dL
\item Borderline high: Total cholesterol 200-239 mg/dL, Non-HDL cholesterol 130-159 mg/dL, HDL cholesterol ≥ 40 mg/dL, and LDL cholesterol 100-159 mg/dL
\item High: Total cholesterol ≥ 240 mg/dL, Non-HDL cholesterol ≥ 160 mg/dL, HDL cholesterol < 40 mg/dL, and LDL cholesterol ≥ 160 mg/dL
}
}
}
